backend/requirements.txt
groq==0.15.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic>=2.0.0,<2.10
python-dotenv==1.0.1
httpx>=0.25.0
supabase>=2.0.0
